+++ packages/building-rollup/CHANGELOG.md
@@ -3,15 +3,29 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-# [2.0.0](https://github.com/open-wc/open-wc/compare/@open-wc/building-rollup@1.1.2...@open-wc/building-rollup@2.0.0) (2020-04-18)
+# [1.2.0](https://github.com/open-wc/open-wc/compare/@open-wc/building-rollup@1.1.4...@open-wc/building-rollup@1.2.0) (2020-04-24)
 
 
-* Chore/update prettier (#1525) ([c55e339](https://github.com/open-wc/open-wc/commit/c55e339343154aefa6665837de1b6ae013f7fe5a)), closes [#1525](https://github.com/open-wc/open-wc/issues/1525)
+### Features
+
+* **building-rollup:** support overriding built-in babel plugins ([ecd294e](https://github.com/open-wc/open-wc/commit/ecd294efad9648db5a0f0c8ed84833e621b757aa))
+* **building-rollup:** use new preserveEntrySignatures option ([ee31fba](https://github.com/open-wc/open-wc/commit/ee31fba7de938362e38ee3f75999c3afdfbb2301))
 
 
-### BREAKING CHANGES
 
-* see prettier release notes
+
+
+## [1.1.4](https://github.com/open-wc/open-wc/compare/@open-wc/building-rollup@1.1.3...@open-wc/building-rollup@1.1.4) (2020-04-21)
+
+**Note:** Version bump only for package @open-wc/building-rollup
+
+
+
+
+
+## [1.1.3](https://github.com/open-wc/open-wc/compare/@open-wc/building-rollup@1.1.2...@open-wc/building-rollup@1.1.3) (2020-04-20)
+
+**Note:** Version bump only for package @open-wc/building-rollup