+++ packages/rollup-plugin-html/CHANGELOG.md
@@ -3,7 +3,7 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-## [1.0.1](https://github.com/open-wc/open-wc/compare/@open-wc/rollup-plugin-html@1.0.0...@open-wc/rollup-plugin-html@1.0.1) (2020-04-18)
+## [1.0.2](https://github.com/open-wc/open-wc/compare/@open-wc/rollup-plugin-html@1.0.0...@open-wc/rollup-plugin-html@1.0.2) (2020-04-20)
 
 **Note:** Version bump only for package @open-wc/rollup-plugin-html